Re-wrote nexafs broadening
